refactor: extract shared validation logic into reusable utility
Created a centralized validation utility to eliminate code duplication
between AddPromptForm and EditPromptForm, improving maintainability and
ensuring consistent validation behavior across all prompt forms.

Changes:
- Created src/utils/validation.ts with validatePromptFields function
- Exported FieldErrors interface and validation function from utils/index.ts
- Updated AddPromptForm to use shared validation (removed 30 lines)
- Updated EditPromptForm to use shared validation (removed 32 lines)
- Maintained all logging context including optional promptId parameter

Benefits:
- Eliminates 62 lines of duplicated validation code
- Single source of truth for prompt validation rules
- Easier to maintain and update validation logic
- Consistent error messages across all forms
- Improved type safety with shared FieldErrors interface

Testing:
- All 952 tests passing (28 for AddPromptForm, 18 for EditPromptForm)
- No ESLint errors
- All functionality preserved with identical validation behavior
