+++ src/utils/index.ts
@@ -8,3 +8,4 @@ export {
   decodeObjectFromBase64UrlSafe,
   isValidBase64UrlSafe
 } from './base64';
+export { validatePromptFields, type FieldErrors } from './validation';
